+++ docs/src/components/LandingLayout.tsx
@@ -3,10 +3,15 @@ import { styled } from "linaria/react";
 
 import SEO from "./SEO";
 import LandingFooter from "./LandingFooter";
-import Banner, { DefaultBanner } from "./ui/Banner";
-import { media, ThemeProvider, tm, tmDark, tmSelectors } from "../themes";
-import { DefaultBannerProps } from "./ui/types";
-import { bannerContent, menuItemsList, socialsItems } from "../config";
+import {
+  headerTotalHeight,
+  media,
+  ThemeProvider,
+  tm,
+  tmDark,
+  tmSelectors,
+} from "../themes";
+import { menuItemsList, socialsItems } from "../config";
 import GDPRNotice from "./GDPRNotice";
 import DocsNavigation from "./DocsNavigation";
 import {
@@ -16,6 +21,7 @@ import {
 } from "./DocumentationLayout";
 import MobileSidebarMenu from "./MobileSidebarMenu";
 import { IDocumentationSidebarStructure, ISeo } from "./types";
+import AlphaBanner from "./ui/AlphaBanner";
 
 const Container = styled.div`
   position: relative;
@@ -39,7 +45,7 @@ const Container = styled.div`
 
 const Main = styled.main`
   overflow-x: hidden;
-  padding-top: 136px;
+  padding-top: ${headerTotalHeight};
   flex: 1 1 auto;
   display: flex;
   flex-direction: column;
@@ -84,16 +90,11 @@ const LandingLayout = ({ children, seo, sidebarLayout }: Props) => {
     <ThemeProvider>
       <Container className="landing">
         <Header>
-          <Banner
-            content={bannerContent}
-            renderContent={({ content }: DefaultBannerProps) => (
-              <DefaultBanner content={content} />
-            )}
-          />
           <DocsNavigation
             isSidebarOpen={isSidebarOpen}
             onSidebarOpen={setIsSidebarOpen}
           />
+          <AlphaBanner />
         </Header>
 
         <SEO seo={seo} />