+++ packages/hardhat-core/src/common/bigInt.ts
@@ -1,6 +1,6 @@
 import type { BigNumber as EthersBigNumberType } from "ethers-v5";
 // eslint-disable-next-line import/no-extraneous-dependencies
-import type { BigNumber as BigNumberJsType } from "bignumber.js";
+import type { default as BigNumberJsType } from "bignumber.js";
 // eslint-disable-next-line import/no-extraneous-dependencies
 import type { default as BNType } from "bn.js";