+++ packages/hardhat-ignition-core/CHANGELOG.md
@@ -1,8 +1,17 @@
 # Changelog
 
-All notable changes to this project will be documented in this file.
+## 0.15.11
 
-The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
+### Patch Changes
+
+- 23280b8: Resolve all dependencies when using submodules in `after`
+- 1a6203d: Add arbitrum sepolia to chain config
+
+## 0.15.10
+
+### Patch Changes
+
+- d96c003: Fix for bug when we fail to save transaction hash
 
 ## 0.15.9 - 2024-12-18